Combine StreamAvailableLocales.java and bug4122700.java
